requirements.txt
numpy>=1.21
scipy>=1.7
tqdm>=4.26
matplotlib>=3.5
scikit-learn>=1.0
casadi>=3.5
pyquaternion>=0.9
joblib>=1.1
pandas>=1.3
PyYAML>=6.0
pycryptodomex>=3.14
gnupg>=2.3
rospkg>=1.3
torch>=1.9
tikzplotlib>=0.9

-i https://rospypi.github.io/simple/
rospy